backend working - move file read logic (StorageService => FileService)
