+++ donetop-main/src/main/java/com/donetop/main/api/file/FileAPIController.java
@@ -1,9 +1,7 @@
 package com.donetop.main.api.file;
 
-import com.donetop.domain.entity.folder.Folder;
 import com.donetop.dto.file.FileDTO;
 import com.donetop.main.service.file.FileService;
-import com.donetop.common.service.storage.StorageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.MediaType;
@@ -24,14 +22,12 @@ public class FileAPIController {
 
 	private final FileService fileService;
 
-	private final StorageService<Folder> storageService;
-
 	@GetMapping(value = SINGULAR + "/{id}")
 	public ResponseEntity<InputStreamResource> get(@PathVariable("id") final long id) {
 		final FileDTO fileDTO = fileService.getFile(id);
 		return ResponseEntity.ok()
 			.contentType(MediaType.valueOf(fileDTO.getMimeType()))
-			.body(storageService.read(fileDTO.getPath()));
+			.body(fileService.read(fileDTO.getPath()));
 	}
 
 }
