+++ donetop-main/src/main/java/com/donetop/main/service/file/FileService.java
@@ -1,9 +1,11 @@
 package com.donetop.main.service.file;
 
 import com.donetop.dto.file.FileDTO;
+import org.springframework.core.io.InputStreamResource;
 
 public interface FileService {
 
 	FileDTO getFile(long id);
 
+	InputStreamResource read(String path);
 }
