+++ donetop-oss/src/main/java/com/donetop/oss/service/file/FileService.java
@@ -1,9 +1,12 @@
 package com.donetop.oss.service.file;
 
 import com.donetop.dto.file.FileDTO;
+import org.springframework.core.io.InputStreamResource;
 
 public interface FileService {
 
 	FileDTO getFile(long id);
+
+	InputStreamResource read(String path);
 
 }
