+++ donetop-oss/src/main/java/com/donetop/oss/service/file/FileServiceImpl.java
@@ -3,8 +3,14 @@ package com.donetop.oss.service.file;
 import com.donetop.dto.file.FileDTO;
 import com.donetop.repository.file.FileRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.core.io.InputStreamResource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Objects;
 
 import static com.donetop.common.api.Message.UNKNOWN_FILE_WITH_ARGUMENTS;
 
@@ -20,5 +26,14 @@ public class FileServiceImpl implements FileService {
 		return fileRepository.findById(id)
 			.orElseThrow(() -> new IllegalStateException(String.format(UNKNOWN_FILE_WITH_ARGUMENTS, id)))
 			.toDTO();
+	}
+
+	@Override
+	public InputStreamResource read(final String path) {
+		try {
+			return new InputStreamResource(Files.newInputStream(Path.of(Objects.requireNonNull(path))));
+		} catch (IOException e) {
+			throw new RuntimeException(e);
+		}
 	}
 }
